refactor: dto 패키지 삭제 후 적절한 도메인 패키지로 이동
