+++ src/main/java/chess/domain/board/BoardOutput.java
@@ -1,4 +1,4 @@
-package chess.dto;
+package chess.domain.board;
 
 import chess.domain.position.Square;
 import chess.domain.piece.Piece;
