+++ src/main/java/chess/domain/position/SquareDifferent.java
@@ -1,4 +1,4 @@
-package chess.dto;
+package chess.domain.position;
 
 public record SquareDifferent(int rankDiff, int fileDiff) {
 }
