+++ src/test/java/chess/domain/DirectionTest.java
@@ -1,9 +1,10 @@
 package chess.domain;
 
+import chess.domain.position.Direction;
 import chess.domain.position.File;
 import chess.domain.position.Rank;
 import chess.domain.position.Square;
-import chess.dto.SquareDifferent;
+import chess.domain.position.SquareDifferent;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
