+++ src/test/java/chess/domain/board/BoardTest.java
@@ -6,7 +6,6 @@ import chess.domain.piece.PieceType;
 import chess.domain.position.File;
 import chess.domain.position.Rank;
 import chess.domain.position.Square;
-import chess.dto.BoardOutput;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
